backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
ddgs>=9.0.0
groq>=0.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
pydantic>=2.8.2
python-dotenv>=1.0.1